+++ examples/PYCO_ETH/DIO/DIO_loopback_port.py
@@ -5,7 +5,11 @@ This example shows how to use DIO loopback with port
 It involves using DO pins to send signals and DI pins to receive signals on a single device, commonly known as "loopback".
 It performs the operation of writing to a DO pin and reading from a DI pin.
 
-Copyright (c) 2023-2024 WPC Systems Ltd. All rights reserved.
+For other examples please check:
+    https://github.com/WPC-Systems-Ltd/WPC_Stand-alone_Python_release/tree/main/examples
+
+Copyright (c) 2024 WPC Systems Ltd.
+All rights reserved.
 '''
 
 ## WPC
@@ -15,9 +19,9 @@ import pywpc
 pin_state = [1, 0, 1, 0, 1, 0, 1]
 
 ## Write port to high or low
-status = pywpc.DO_writePort(pin_state)
-print("DO_writePort status: ", status)
+status = pywpc.DO_writeAllChannels(pin_state)
+print("DO_writeAllChannels status: ", status)
 
 ## Read DI port state
-state_list = pywpc.DI_readPort()
-print("DI_readPort: ", state_list)
+state_list = pywpc.DO_readAllChannels()
+print("DO_readAllChannels: ", state_list)