+++ examples/PYCO_WIFI/TCP_client/TCP_client_write_U8_list.py
@@ -3,7 +3,11 @@ TCP_client - TCP_client_writeU8List.py.
 
 This example demonstrates how to write u8 list to server
 
-Copyright (c) 2023-2024 WPC Systems Ltd. All rights reserved.
+For other examples please check:
+    https://github.com/WPC-Systems-Ltd/WPC_Stand-alone_Python_release/tree/main/examples
+
+Copyright (c) 2024 WPC Systems Ltd.
+All rights reserved.
 '''
 
 ## WPC