+++ examples/PYCO_WIFI_MINI/Modbus_slave/write_data_with_register.py
@@ -5,7 +5,11 @@ This example demonstrates how to write data into slave
 
 The range of input registers is from 41000 to 41499.
 
-Copyright (c) 2023-2024 WPC Systems Ltd. All rights reserved.
+For other examples please check:
+    https://github.com/WPC-Systems-Ltd/WPC_Stand-alone_Python_release/tree/main/examples
+
+Copyright (c) 2024 WPC Systems Ltd.
+All rights reserved.
 '''
 
 ## WPC