+++ Changelog.md
@@ -1,6 +1,16 @@
 WPC Stand-alone_python_release changelog
 =======================================
 
+v0.7.6 Date: 2024/06/07, Developer: Chunglee_people
+---------------------------------------------------
+### Added
+- Add WPC error code link into API return function
+- example code: `get started.py` and `get network info.py`
+- API: Sys_getSerialNumber, Sys_getIP, Sys_getSubmask, Sys_getGateway
+
+### Fixed
+- API: Sys_reboot API
+
 v0.7.0 Date: 2024/06/03, Developer: Chunglee_people
 ---------------------------------------------------
 ### Added